changed files to js files
