+++ quality_hub/src/App.js
@@ -13,7 +13,7 @@ import StudentForm from './components/StudentForm';
 function App() {
   return (
     <>
-      <Route exact path='/' component={Landing} />
+      <Route exact path='/' component={Dashboard} />
       <Route path='/login/' component={LoginForm} />
       <Route path='/register' component={SignUpForm} />
       <Route path='/interviewer' component={InterviewerForm} />
