+++ quality_hub/src/components/Landing.js
