+++ quality_hub/src/components/devices.js
