+++ quality_hub/src/state/actions/actionCreators.js
@@ -1,10 +1,12 @@
 import * as types from './actionTypes';
 import axios from 'axios';
 
-export const login = (url, props, values) => dispatch => {
+const url = 'https://dev-coach-staging.herokuapp.com/';
+
+export const login = (props, values) => dispatch => {
   dispatch({ type: types.LOGIN_START });
   axios
-    .post(url, values)
+    .post(`${url}user/login`, values)
     .then(res => {
       dispatch({ type: types.LOGIN_SUCCESSFUL });
       localStorage.setItem('token', res.data.token);
@@ -16,12 +18,13 @@ export const login = (url, props, values) => dispatch => {
     });
 };
 
-export const register = (url, props, values) => dispatch => {
+export const register = (props, values) => dispatch => {
   dispatch({ type: types.SIGN_UP });
   dispatch({ type: types.LOGIN_START });
   axios
-    .post(url, values)
+    .post(`${url}user/register`, values)
     .then(res => {
+      debugger
       dispatch({ type: types.SIGN_UP_SUCCESSFUL });
       localStorage.setItem('user', JSON.stringify(res.data));
       dispatch({ type: types.LOGIN_SUCCESSFUL });
@@ -30,6 +33,7 @@ export const register = (url, props, values) => dispatch => {
       props.history.push('/dashboard');
     })
     .catch(err => {
+      debugger
       dispatch({ type: types.SIGN_UP_ERROR, payload: err });
     });
 };
